refactor :: file structer
파일 구조 전반적인 부분 수정,
하드코딩 되던 hook파일 분리, ui 또한 commponent로 분리 작업 수행중
